pyproject.toml
[tool.poetry]
name = "wselfedu-test"
version = "0.1.0"
description = ""
authors = ["Sergei Mikurov <454004@mail.ru>"]
readme = "README.md"
packages = [{include = "wselfedu_test"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest-playwright = "^0.4.4"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
